+++ example/TestLinkExampleGenericApi.py
@@ -73,7 +73,8 @@ projNr=len(myTestLink.getProjects())+1
 
 NEWPROJECT="PROJECT_API_GENERIC-%i" % projNr
 NEWPREFIX="GPROAPI%i" % projNr
-NEWTESTPLAN="TestPlan_API_GENERIC %i" % projNr
+NEWTESTPLAN_A="TestPlan_API_GENERIC A"
+NEWTESTPLAN_B="TestPlan_API_GENERIC B"
 NEWPLATFORM_A='Big Bird %i' % projNr
 NEWPLATFORM_B='Small Bird'
 NEWPLATFORM_C='Ugly Bird'
@@ -113,19 +114,29 @@ print ""
 
 # # Creates the project
 newProject = myTestLink.createTestProject(NEWPROJECT, NEWPREFIX, 
-    notes='This is a Project created with the Generic API', active=1, public=1,
+    notes='Example created with Python API class %s' % NEWBUILD, 
+    active=1, public=1,
     options={'requirementsEnabled' : 1, 'testPriorityEnabled' : 1, 
              'automationEnabled' : 1,  'inventoryEnabled' : 1})
 print "createTestProject", newProject
 newProjectID = newProject[0]['id'] 
 print "New Project '%s' - id: %s" % (NEWPROJECT,newProjectID)
  
-# Creates the test plan
-newTestPlan = myTestLink.createTestPlan(NEWTESTPLAN, NEWPROJECT,
-            notes='New TestPlan created with the Generic API',active=1, public=1)    
+# Create test plan A  - uses platforms
+newTestPlan = myTestLink.createTestPlan(NEWTESTPLAN_A, NEWPROJECT,
+            notes='New TestPlan created with the Generic API - uses platforms.',
+            active=1, public=1)    
 print "createTestPlan", newTestPlan
-newTestPlanID = newTestPlan[0]['id'] 
-print "New Test Plan '%s' - id: %s" % (NEWTESTPLAN,newTestPlanID)
+newTestPlanID_A = newTestPlan[0]['id'] 
+print "New Test Plan '%s' - id: %s" % (NEWTESTPLAN_A,newTestPlanID_A)
+
+# Create test plan B  - uses no platforms
+newTestPlan = myTestLink.createTestPlan(NEWTESTPLAN_B, NEWPROJECT,
+            notes='New TestPlan created with the Generic API - uses no platforms.',
+            active=1, public=1)    
+print "createTestPlan", newTestPlan
+newTestPlanID_B = newTestPlan[0]['id'] 
+print "New Test Plan '%s' - id: %s" % (NEWTESTPLAN_B,newTestPlanID_B)
  
 # Create platform 'Big Bird x' 
 newPlatForm = myTestLink.createPlatform(NEWPROJECT, NEWPLATFORM_A, 
@@ -133,7 +144,7 @@ newPlatForm = myTestLink.createPlatform(NEWPROJECT, NEWPLATFORM_A,
 print "createPlatform", newPlatForm
 newPlatFormID_A = newPlatForm['id']
 # Add Platform  'Big Bird x' to platform 
-response = myTestLink.addPlatformToTestPlan(newTestPlanID, NEWPLATFORM_A) 
+response = myTestLink.addPlatformToTestPlan(newTestPlanID_A, NEWPLATFORM_A) 
 print "addPlatformToTestPlan", response
 
 # Create platform 'Small Bird'
@@ -142,7 +153,7 @@ newPlatForm = myTestLink.createPlatform(NEWPROJECT, NEWPLATFORM_B,
 print "createPlatform", newPlatForm
 newPlatFormID_B = newPlatForm['id']
 # Add Platform  'Small Bird' to platform 
-response = myTestLink.addPlatformToTestPlan(newTestPlanID, NEWPLATFORM_B) 
+response = myTestLink.addPlatformToTestPlan(newTestPlanID_A, NEWPLATFORM_B) 
 print "addPlatformToTestPlan", response
 
 # Create platform 'Ugly Bird'
@@ -151,7 +162,7 @@ newPlatForm = myTestLink.createPlatform(NEWPROJECT, NEWPLATFORM_C,
 print "createPlatform", newPlatForm
 newPlatFormID_C = newPlatForm['id']
 # Add Platform  'Ugly Bird' to platform 
-response = myTestLink.addPlatformToTestPlan(newTestPlanID, NEWPLATFORM_C) 
+response = myTestLink.addPlatformToTestPlan(newTestPlanID_A, NEWPLATFORM_C) 
 print "addPlatformToTestPlan", response
 
 # looking, which platforms exists
@@ -228,38 +239,38 @@ newTestCaseID_B = newTestCase[0]['id']
 print "New Test Case '%s' - id: %s" % (NEWTESTCASE_B, newTestCaseID_B)
   
 # Add  test cases to test plan - we need the full external id !
+# for every test case version 1 is used
+tc_version=1
 # TC AA should be tested with platforms 'Big Bird'+'Small Bird'
 tc_aa_full_ext_id = myTestLink.getTestCase(testcaseid=newTestCaseID_AA)[0]['full_tc_external_id']
-response = myTestLink.callServerWithPosArgs('addTestCaseToTestPlan', 
-                devKey=myTestLink.devKey, testprojectid=newProjectID, 
-                testplanid=newTestPlanID, testcaseexternalid=tc_aa_full_ext_id,
-                platformid=newPlatFormID_A,version=1)
+response = myTestLink.addTestCaseToTestPlan(newProjectID, newTestPlanID_A, 
+                    tc_aa_full_ext_id, tc_version, platformid=newPlatFormID_A)
 print "addTestCaseToTestPlan", response
 tc_aa_full_ext_id = myTestLink.getTestCase(testcaseid=newTestCaseID_AA)[0]['full_tc_external_id']
-response = myTestLink.callServerWithPosArgs('addTestCaseToTestPlan', 
-                devKey=myTestLink.devKey, testprojectid=newProjectID, 
-                testplanid=newTestPlanID, testcaseexternalid=tc_aa_full_ext_id,
-                platformid=newPlatFormID_B,version=1)
+response = myTestLink.addTestCaseToTestPlan(newProjectID, newTestPlanID_A, 
+                    tc_aa_full_ext_id, tc_version, platformid=newPlatFormID_B)
 print "addTestCaseToTestPlan", response
 # TC B should be tested with platform 'Small Bird'
 tc_b_full_ext_id = myTestLink.getTestCase(testcaseid=newTestCaseID_B)[0]['full_tc_external_id']
-response = myTestLink.callServerWithPosArgs('addTestCaseToTestPlan', 
-                devKey=myTestLink.devKey, testprojectid=newProjectID, 
-                testplanid=newTestPlanID, testcaseexternalid=tc_b_full_ext_id,
-                platformid=newPlatFormID_B,version=1)
+response = myTestLink.addTestCaseToTestPlan(newProjectID, newTestPlanID_A, 
+                    tc_b_full_ext_id, tc_version, platformid=newPlatFormID_B)
+print "addTestCaseToTestPlan", response
+# In test plan B TC B should be tested without  platform 
+response = myTestLink.addTestCaseToTestPlan(newProjectID, newTestPlanID_B, 
+                                            tc_b_full_ext_id, tc_version)
 print "addTestCaseToTestPlan", response
 
 # # Try to Remove Platform  'Big Bird' from platform 
-# response = myTestLink.removePlatformFromTestPlan(newTestPlanID, NEWPLATFORM_C) 
+# response = myTestLink.removePlatformFromTestPlan(newTestPlanID_A, NEWPLATFORM_C) 
 # print "removePlatformFromTestPlan", response
 
 # Remove Platform  'Ugly Bird' from platform 
-response = myTestLink.removePlatformFromTestPlan(newTestPlanID, NEWPLATFORM_C) 
+response = myTestLink.removePlatformFromTestPlan(newTestPlanID_A, NEWPLATFORM_C) 
 print "removePlatformFromTestPlan", response
 
 
 # -- Create Build
-newBuild = myTestLink.createBuild(newTestPlanID, NEWBUILD, 
+newBuild = myTestLink.createBuild(newTestPlanID_A, NEWBUILD, 
                                   buildnotes='Notes for the Build')
 print "createBuild", newBuild
 newBuildID = newBuild[0]['id'] 
@@ -267,20 +278,20 @@ print "New Build '%s' - id: %s" % (NEWBUILD, newBuildID)
   
 # report Test Case Results for platform 'Big Bird'
 # TC_AA failed, build should be guessed, TC identified with external id
-newResult = myTestLink.reportTCResult(newTestPlanID, 'f', guess=True,
+newResult = myTestLink.reportTCResult(newTestPlanID_A, 'f', guess=True,
                                       testcaseexternalid=tc_aa_full_ext_id,
                                       platformname=NEWPLATFORM_A)
 print "reportTCResult", newResult
 newResultID_AA = newResult[0]['id']
 # report Test Case Results for platform 'Small Bird'
 # TC_AA passed, build should be guessed, TC identified with external id
-newResult = myTestLink.reportTCResult(newTestPlanID, 'p', guess=True,
+newResult = myTestLink.reportTCResult(newTestPlanID_A, 'p', guess=True,
                                       testcaseexternalid=tc_aa_full_ext_id,
                                       platformid=newPlatFormID_B)
 print "reportTCResult", newResult
 newResultID_AA_p = newResult[0]['id']
 # TC_B passed, explicit build and some notes , TC identified with internal id
-newResult = myTestLink.reportTCResult(newTestPlanID, 'p', 
+newResult = myTestLink.reportTCResult(newTestPlanID_A, 'p', 
                 buildid=newBuildID, testcaseid=newTestCaseID_B, 
                 platformname=NEWPLATFORM_B, notes="first try")
 print "reportTCResult", newResult
@@ -310,20 +321,20 @@ response = myTestLink.getFirstLevelTestSuitesForTestProject(newProjectID)
 print "getFirstLevelTestSuitesForTestProject", response
 
 # get information - TestPlan
-response = myTestLink.getTestPlanByName(NEWPROJECT, NEWTESTPLAN)
+response = myTestLink.getTestPlanByName(NEWPROJECT, NEWTESTPLAN_A)
 print "getTestPlanByName", response
-response = myTestLink.getTotalsForTestPlan(newTestPlanID)
+response = myTestLink.getTotalsForTestPlan(newTestPlanID_A)
 print "getTotalsForTestPlan", response
-response = myTestLink.getBuildsForTestPlan(newTestPlanID)
+response = myTestLink.getBuildsForTestPlan(newTestPlanID_A)
 print "getBuildsForTestPlan", response
-response = myTestLink.getLatestBuildForTestPlan(newTestPlanID)
+response = myTestLink.getLatestBuildForTestPlan(newTestPlanID_A)
 print "getLatestBuildForTestPlan", response
-response = myTestLink.getTestPlanPlatforms(newTestPlanID)
+response = myTestLink.getTestPlanPlatforms(newTestPlanID_A)
 print "getTestPlanPlatforms", response
-response = myTestLink.getTestSuitesForTestPlan(newTestPlanID)
+response = myTestLink.getTestSuitesForTestPlan(newTestPlanID_A)
 print "getTestSuitesForTestPlan", response
 # get failed Testcases 
-response = myTestLink.getTestCasesForTestPlan(newTestPlanID, executestatus='f')
+response = myTestLink.getTestCasesForTestPlan(newTestPlanID_A, executestatus='f')
 print "getTestCasesForTestPlan", response
 
 # get information - TestSuite
@@ -346,10 +357,10 @@ tcpathname = '::'.join([NEWPROJECT, NEWTESTSUITE_A, NEWTESTSUITE_AA, NEWTESTCASE
 response = myTestLink.getTestCaseIDByName('unknown', testcasepathname=tcpathname)
 print "getTestCaseIDByName", response
 # get execution result
-response = myTestLink.getLastExecutionResult(newTestPlanID, 
+response = myTestLink.getLastExecutionResult(newTestPlanID_A, 
                                              testcaseexternalid=tc_aa_full_ext_id)
 print "getLastExecutionResult", response
-response = myTestLink.getLastExecutionResult(newTestPlanID, 
+response = myTestLink.getLastExecutionResult(newTestPlanID_A, 
                                              testcaseid=newTestCaseID_B)
 print "getLastExecutionResult", response