+++ CHANGES.rst
@@ -1,6 +1,12 @@
 Changes in TestLink-API-Python-client Source Distribution
 =========================================================
 
+TestLink-API-Python-client UNDER DEVELOP v0.4.7
+-----------------------------------------------
+
+implement missing 1.9.8 api methods - TestCase
+- addTestCaseToTestPlan 
+
 TestLink-API-Python-client release notes v0.4.6 (Dec. 2013)
 -----------------------------------------------------------